[REFACTOR] #26: 엔티티 생성자에 추가된 칼럼도 포함하도록 수정
