--- src/main/java/com/deark/be/event/domain/ThumbnailSource.java
@@ -1,5 +1,0 @@
-package com.deark.be.event.domain;
-
-public enum ThumbnailSource {
-    DESIGN, STORE
-}
