+++ src/main/java/com/deark/be/event/domain/type/ThumbnailSource.java
@@ -1,0 +1,7 @@
+package com.deark.be.event.domain.type;
+
+public enum ThumbnailSource {
+    DESIGN,
+    STORE,
+    ;
+}
